python_scripts/requirements.txt
functions-framework
google-cloud-translate
google-genai
firebase-admin
requests